pyproject.toml
[tool.poetry]
name = "draft polling"
version = "0.1.0"
description = ""
authors = ["Giorgos Gagrinas <gagrinas.g@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
apscheduler = "^3.10.4"
uvicorn = "^0.24.0.post1"
fastapi-utils = "^0.2.1"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
